+++ fava/static/javascript/lib/events.ts
@@ -51,3 +51,54 @@ export class Events<T = string> {
     });
   }
 }
+
+/**
+ * Execute the callback of the event of given type is fired on something
+ * matching selector.
+ *
+ * @param element - The ancestor element that the listener will be attached to.
+ * @param type - The event type to listen to.
+ * @param selector - The DOM selector to match.
+ * @param callback - The event listener to execute on a match.
+ */
+export function delegate<T extends Event, C extends Element>(
+  element: Element | Document,
+  type: string,
+  selector: string,
+  callback: (e: T, c: C) => void
+): void {
+  element.addEventListener(type, (event) => {
+    let { target } = event;
+    if (!(target instanceof Node)) {
+      return;
+    }
+    if (!(target instanceof Element)) {
+      target = target.parentNode;
+    }
+    if (target instanceof Element) {
+      const closest = target.closest(selector);
+      if (closest) {
+        callback(event as T, closest as C);
+      }
+    }
+  });
+}
+
+/**
+ * Bind an event to element, only run the callback once.
+ * @param element - The element to attach the listener to.
+ * @param event - The event type.
+ * @param callback - The event listener.
+ */
+export function once(
+  element: EventTarget,
+  event: string,
+  callback: (ev: Event) => void
+): void {
+  function runOnce(ev: Event): void {
+    element.removeEventListener(event, runOnce);
+    callback.apply(element, [ev]);
+  }
+
+  element.addEventListener(event, runOnce);
+}
